expose hdri environmental light
with pathtracer there is a new checkbox to load an environmental
texture, similar to backplate texture. When loaded it adds
itself within an hdri light, to the scene. Note that the
default ambient light is too bright, so the environmental light
is not visible until you use the light editor to make it dimmer.

Minor cleanup to the brand new light editor to accomodate
including getting rid of remaining vestiges of the global light.