reorganize the GUI and deemphasize tutorial features
The choice of demo scene is moved to the File menu and
redundant import choices are removed. The application also
starts with an empty scene now. That
seems a bit more professional to me than the old default.

The floating panel controls are moved to the Edit menu and
organized into catagories there. The precanned backplate
is removed because, without contols over colors, I don't
think it will be used in production.

The preferences entry is removed too since the one item in
it was not saved across sessions and thus wasn't really a
preference per se.