use a NodePtr instead of Node* in Search visitor
originally had `Node *` because the visitor `operator()` API requires
the first argument to be a `Node &`. Since we can't have a vector of
references, I chose to use a raw pointer to point back to search
results. BUT this leads to dangling pointers in the search widget if you
clear the scene (e.g. load demo scene, search for 'xfm' in Transform
editor, clear scene with Transform editor open, kaboom).

However, `Node` does have a public method to return a `NodePtr`,
`nodeAs<T>()`. In this case, we can use `nodeAs<Node>()`. Using the
shared pointer will allow us to safely check if a result is still valid
and erase it as needed. Ideally we should add a `Node` method that
simply returns a `std::make_shared(this)` since it already inherits from
`std::enable_shared_from_this`.