fixup studio python mpi
In the cleanup we moved all of the MPIrank calls into sgMPI state,
but neglected to provide a way for the Python interface to initialize
it. Thus the generators and pvd reader stopped divying up the data
when called from Python.

This change lets python initialize MPI and let sgMPI know about
it to fix that.