remove 32 bit limit on slices generator number of triangles
This adds a long int node type along with python wrapping and a
and a rudimentary imgui widget for it. At this point the widget
treats the long as a normal int.

The studioDistBench1 script is changed to use the long node
for the number of requested triangles. As a bonus it is also
changed to print out the actual number of generated triangles
on every MPI rank.

Finally the waveletslices generator is changed to accept either
int, uint, or long nodes, because python3's int stores long
but give whichever one fits to c++ on the fly.

Taken together python scripts should now be able to overflow
ospray's 32bit mesh index on each MPI rank if we ask them too.